mint.cpp

+++ docs/teamnote/Misc/Mint/mint.cpp
@@ -17,7 +17,7 @@ struct mint
     operator int(void) const { return x; }
 };
 
-mint mpow(mint a, int x)
+mint mpow(mint a, ll x)
 {
     mint ret=1;
     while(x)
